scripts/requirements-audit.txt
beautifulsoup4>=4.12.3
requests>=2.32.3
Pillow>=10.4.0
ImageHash>=4.3.1
openpyxl>=3.1.5
